requirements.txt
cluster==1.4.1.post3
gensim==4.1.2
javalang==0.13.0
numpy==1.19.5
pandas==1.3.4
process_data==0.3
pycparser==2.19
pyenchant==3.2.2
scikit_learn==1.0.2
scipy==1.4.1
torch==1.7.0
tqdm==4.42.1
zss==1.2.0
